refactor: Modal의 상위폴더 GenreModal 추가
