+++ music/src/Pages/Result/Hip.js
@@ -12,16 +12,16 @@ import {
   useDisclosure,
 } from '@chakra-ui/react';
 
+import CommonModal from '../../Components/modal/GenreModal/Modal';
 import {
   DrillHipModalBody,
   TrapHipModalBody,
   AlterHipModalBody,
   RageHipModalBody,
   BoomHipModalBody,
-} from '../../Components/modal/ModalContent';
+} from '../../Components/modal/GenreModal/ModalContent';
 
 import Share from '../../Components/Share/Share';
-import CommonModal from '../../Components/modal/Modal';
 import LazyImage from '../../Components/Image/LazyImage';
 import './Result.css';
